+++ scripts/i18n/extract.sh
@@ -5,8 +5,8 @@
 for d in 'web' 'api'; 
 do
     pushd $d >/dev/null
-    /chia-blockchain/venv/bin/pybabel extract -F babel.cfg -k _l -o messages.pot .
-    /chia-blockchain/venv/bin/pybabel update -i messages.pot -d ./translations
+    /chia-blockchain/venv/bin/pybabel --verbose extract -F babel.cfg -k _l -o messages.pot .
+    /chia-blockchain/venv/bin/pybabel --verbose update -i messages.pot -d ./translations
     chmod -R 777 $PWD/translations
     popd >/dev/null
 done