+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file. The format 
 ## [Unreleased]
 ### Added
  - Support for running a blockchain timelord with environment variable `mode` set to `fullnode,timelord`.
+ - Optional config setting to restart fork fullnodes if they consume too much memory.
 ## Changed
+ - On fresh install, optionally, via libtorrent, download a recent [Chia database checkpoint](https://www.chia.net/downloads/).
  - Fixes for Gigahorse GPU plotting including 64 GB RAM mode using the `tmp3` SSD plotting path.
  - Fix for Gigahorse Alerts, please reset earlier broken settings file from [this one](https://github.com/guydavis/machinaris/blob/develop/config/chiadog/gigahorse.sample.yaml).
  - Fix for container stop signal to cleanly shutdown forks too, as was already happening for Chia.
@@ -16,6 +18,7 @@ All notable changes to this project will be documented in this file. The format 
  - [MMX](https://github.com/madMAx43v3r/mmx-node/releases/tag/v0.9.14) to (v0.9.14) on `testnet9`.  Supporting the latest AMD GPU driver.
  - [One](https://github.com/xone-network/one-blockchain/releases/tag/1.7.0) to v1.7.0
 ## Notes
+ - Bladebit has regressed on the ability to `ramplot` as they try to match support by Gigahorse for GPU plotting. Awaiting a fix...
  - Support for new blockchains and tools DOES NOT imply my endorsement for them.  *Only run those you are comfortable with.*
 
 ## [0.8.7] - 2023-02-16